+++ spec/valueexp/IValExpr.java
@@ -4,6 +4,6 @@ import java.util.Optional;
 
 public interface IValExpr {
 
-	Optional<Object> calculateValueAt(int nt);
+	Optional<Object> calculateValueAt(double nt);
 
 }
