+++ spec/valueexp/RandomIntExpr.java
@@ -5,8 +5,8 @@ import java.util.Optional;
 public class RandomIntExpr implements IValExpr {
 
 	@Override
-	public Optional<Object> calculateValueAt(int nt) {
-		return Optional.of(5);
+	public Optional<Object> calculateValueAt(double nt) {
+		return Optional.of(5d);
 	}
 
 }
